requirements.txt
discord.py[voice]
pynacl
yt-dlp
spotipy
python-dotenv